% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KeyFunctions.R
\name{BindMatrices}
\alias{BindMatrices}
\title{Bind Matrices}
\usage{
BindMatrices(cluster_list, correlation_matrix)
}
\arguments{
\item{cluster_list}{A list of clusters to generate adjacency matrices.}

\item{correlation_matrix}{A correlation matrix to align with the adjacency matrix.}
}
\value{
A list containing the combined adjacency matrix and CCCN matrix.
}
\description{
This function binds matrices, aligns them, and prepares adjacency and CCCN matrices.
}
\examples{
BindMatrices(cluster_list, correlation_matrix)
}
