% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KeyFunctions.R
\name{EuclideanDistance}
\alias{EuclideanDistance}
\title{Calculates Euclidean distance and performs t-SNE}
\usage{
EuclideanDistance(ptmtable)
}
\arguments{
\item{ptmtable}{A data frame containing numeric data for post-translational modifications.}
}
\value{
A matrix containing t-SNE coordinates (3D).
}
\description{
This function computes the Euclidean distance matrix from the input dataset,
normalizes it, and applies t-SNE for dimensionality reduction.
}
\examples{
EuclideanDistance(ptmtable)
}
