% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KeyFunctions.R
\name{extract.gene.names}
\alias{extract.gene.names}
\title{Extract Gene Names from Peptide Edge File}
\usage{
extract.gene.names(peptide.edgefile)
}
\arguments{
\item{peptide.edgefile}{A data frame containing peptide edge information.}
}
\value{
A vector of unique gene names.
}
\description{
This function extracts unique gene names from a peptide edge file.
}
\examples{
extract.gene.names(peptide.edgefile)
}
