% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KeyFunctions.R
\name{find_ppi_edges}
\alias{find_ppi_edges}
\title{Find PPI Edges}
\usage{
find_ppi_edges(input_dataset, gmfilename, nodenames)
}
\arguments{
\item{input_dataset}{The input dataset containing experimental data.}

\item{gmfilename}{The filename of the GeneMANIA data.}

\item{nodenames}{A vector of node names.}
}
\value{
A data frame of combined edges from STRINGdb and GeneMANIA.
}
\description{
This function finds protein-protein interaction edges by combining input datasets with STRING and GeneMANIA databases.
}
\examples{
find_ppi_edges("input_data.txt", "gmfilename.txt", nodenames)
}
