% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KeyFunctions.R
\name{genepep.edges.3}
\alias{genepep.edges.3}
\title{Create Gene-Peptide Edges}
\usage{
genepep.edges.3(nodelist, pepkey = ld.key)
}
\arguments{
\item{nodelist}{A vector of node names.}

\item{pepkey}{A data frame containing peptide keys.}
}
\value{
A data frame of peptide edges with weights and edge types.
}
\description{
This function creates peptide edges for a given node list.
}
\examples{
genepep.edges.3(nodelist, pepkey)
}
