% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KeyFunctions.R
\name{GenerateAndConstructptmsNetwork}
\alias{GenerateAndConstructptmsNetwork}
\title{Generate and Construct All PTMs Network}
\usage{
GenerateAndConstructptmsNetwork(
  eu.ptms.list,
  sp.ptms.list,
  sed.ptms.list,
  ptmtable,
  keeplength = 2,
  output_dir = "plots"
)
}
\arguments{
\item{eu.ptms.list}{A list containing all PTMs data for the European dataset.}

\item{sp.ptms.list}{A list containing all PTMs data for the SP dataset.}

\item{sed.ptms.list}{A list containing all PTMs data for the SED dataset.}

\item{ptmtable}{A data frame containing all PTMs data.}

\item{keeplength}{An integer specifying the minimum length of common elements to keep. Default is 2.}

\item{output_dir}{A string specifying the output directory for saving plots. Default is "plots".}
}
\value{
A list containing the updated \code{ptmtable} and data for \code{eu.sp.sed.ptms}.
}
\description{
This function generates and constructs the PTMs network from given data lists and tables.
}
\examples{
GenerateAndConstructptmsNetwork(eu.ptms.list, sp.ptms.list, sed.ptms.list, ptmtable)
}
