% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KeyFunctions.R
\name{get.gene.names.from.peps}
\alias{get.gene.names.from.peps}
\title{Extract Gene Names from Peptide Vector}
\usage{
get.gene.names.from.peps(pepvec, pepsep = "; ")
}
\arguments{
\item{pepvec}{A vector of peptides.}

\item{pepsep}{A string specifying the separator for peptides. Default is "; ".}
}
\value{
A vector of unique gene names.
}
\description{
This function extracts gene names from a given peptide vector.
}
\examples{
get.gene.names.from.peps(c("gene1 peptide1", "gene2 peptide2"))
}
