% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KeyFunctions.R
\name{list.common}
\alias{list.common}
\title{Finds common elements between clusters in two lists.}
\usage{
list.common(list1, list2, keeplength = 3)
}
\arguments{
\item{list1}{A list of clusters.}

\item{list2}{A list of clusters to compare against.}

\item{keeplength}{Minimum size of intersections to keep.}
}
\value{
A list of common clusters.
}
\description{
Finds common elements between clusters in two lists.
}
\examples{
list.common(cluster_list1, cluster_list2, keeplength = 3)
}
