% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KeyFunctions.R
\name{MakeAdjMatrix}
\alias{MakeAdjMatrix}
\title{Create Adjacency Matrix}
\usage{
MakeAdjMatrix(list.element)
}
\arguments{
\item{list.element}{A list of elements to construct the adjacency matrix.}
}
\value{
A square matrix where rows and columns correspond to the input list elements.
}
\description{
This function creates an adjacency matrix for a given list element.
}
\examples{
MakeAdjMatrix(c("A", "B", "C"))
}
