% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KeyFunctions.R
\name{SpearmanDissimilarity}
\alias{SpearmanDissimilarity}
\title{Calculates Spearman dissimilarity and t-SNE from a given dataset}
\usage{
SpearmanDissimilarity(ptmtable)
}
\arguments{
\item{ptmtable}{A data frame containing numeric data for post-translational modifications.}
}
\value{
A matrix containing t-SNE coordinates (3D).
}
\description{
This function computes the Spearman dissimilarity matrix from the input dataset,
processes it for missing values, and performs t-SNE for dimensionality reduction.
}
\examples{
SpearmanDissimilarity(ptmtable)
}
