% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KeyFunctions.R
\name{zero.to.NA}
\alias{zero.to.NA}
\title{Replace Zeros with NA}
\usage{
zero.to.NA(df)
}
\arguments{
\item{df}{A data frame where zeros are to be replaced with NA.}
}
\value{
A data frame with zeros replaced by NA.
}
\description{
This function replaces all zeros in a data frame with NA values.
}
\examples{
zero.to.NA(data.frame(a = c(0, 1), b = c(2, 0)))
}
