% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KeyFunctions.R
\name{MakeClusterList}
\alias{MakeClusterList}
\title{Populates the global enviroment with cluster groupings based on t-SNE data}
\usage{
MakeClusterList(ptmtable, toolong = 3.5)
}
\arguments{
\item{toolong}{A numeric threshold for cluster separation, defaults to 3.5.}

\item{sed_ptms_tsne}{A matrix containing t-SNE coordinates.}

\item{tbl.sc}{A data frame associated with the t-SNE data.}
}
\description{
This function groups t-SNE data points into clusters using a specified threshold
and visualizes the clusters.
}
\examples{
MakeClusterList(sed_ptms_tsne, tbl.sc, toolong =  3.5)
}
