% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KeyFunctions.R
\name{MakeClusterList}
\alias{MakeClusterList}
\title{Creates a list of cluster groupings based on t-SNE data}
\usage{
MakeClusterList(tsnedata, toolong, tbl.sc)
}
\arguments{
\item{tsnedata}{A matrix containing t-SNE coordinates.}

\item{toolong}{A numeric threshold for cluster separation.}

\item{tbl.sc}{A data frame associated with the t-SNE data.}
}
\value{
A list of clusters grouped by proximity.
}
\description{
This function groups t-SNE data points into clusters using a specified threshold
and visualizes the clusters.
}
\examples{
MakeClusterList(tsnedata, 3.5, tbl.sc)
}
